+++ Domain/DTOs/Sales/PagoVentaDTO.cs
@@ -6,9 +6,9 @@ public class PagoVentaDTO
     public long IdVenta { get; set; }
     public long IdMetodoPago { get; set; }
     public DateTime Fecha { get; set; }
-    public double MontoTotal { get; set; }
-    public double MontoRecibido { get; set; }
-    public double MontoCambio { get; set; }
+    public decimal MontoTotal { get; set; }
+    public decimal MontoRecibido { get; set; }
+    public decimal MontoCambio { get; set; }
     public string Observacion { get; set; } = string.Empty;
 
     public VentaDTO? Venta { get; set; }
