+++ Domain/Entities/Sales/VentaDetalle.cs
@@ -8,11 +8,11 @@ public class VentaDetalle : Entity
     public long? IdOrdenVentaDetalle { get; set; }
     public long IdProducto { get; set; }
     public string NombreProducto { get; set; } = string.Empty;
-    public double Cantidad { get; set; }
-    public double PrecioUnitario { get; set; }
-    public double? Descuento { get; set; }
-    public double SubTotal { get; set; }
-    public double Total { get; set; }
+    public decimal Cantidad { get; set; }
+    public decimal PrecioUnitario { get; set; }
+    public decimal? Descuento { get; set; }
+    public decimal SubTotal { get; set; }
+    public decimal Total { get; set; }
 
     public Venta? Venta { get; set; }
     public OrdenVentaDetalle? OrdenVentaDetalle { get; set; }
